기어 구조 레벨에 맟추어 대폭 수정, 메모리 풀 구현
메모리 풀을 만들기는 했는데... 사용은 아직 해보지 않았어요.

+++ DJMAX/Client/CEditorLevel.cpp
@@ -234,7 +234,7 @@ INT_PTR CALLBACK NoteEditProc(HWND hDlg, UINT message, WPARAM wParam, LPARAM lPa
 			}
 
 			hComboBox = GetDlgItem(hDlg, IDC_COMBO_NOTETYPE);
-			selectedIndex = SendMessage(hComboBox, CB_GETCURSEL, 0, 0);
+			selectedIndex = (int)SendMessage(hComboBox, CB_GETCURSEL, 0, 0);
 
 			if (selectedIndex != CB_ERR)
 			{
